+++ test/refiner_flutter_test.dart
@@ -1,8 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:refiner_flutter/refiner_flutter.dart';
-import 'package:refiner_flutter/refiner_flutter_platform_interface.dart';
-import 'package:refiner_flutter/refiner_flutter_method_channel.dart';
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+import 'package:refiner_flutter/src/refiner_flutter_platform_interface.dart';
+import 'package:refiner_flutter/src/refiner_flutter_method_channel.dart';
 
 void main() {
   final RefinerFlutterPlatform initialPlatform = RefinerFlutterPlatform.instance;